Update create and destroy methods to check for role

+++ app/controllers/v1/categories_controller.rb
@@ -5,16 +5,24 @@ class V1::CategoriesController < ApplicationController
   end
 
   def create
-    @category = Category.new(name: params[:name])
-    return render json: @category.errors, status: :unprocessable_entity unless @category.save
+    if @current_user.role == 'admin'
+      @category = Category.new(name: params[:name])
+      return render json: @category.errors, status: :unprocessable_entity unless @category.save
 
-    render json: { status: :created, message: 'Category created successfully.', body: @category }, status: :created
+      render json: { status: :created, message: 'Category created successfully.', body: @category }, status: :created
+    else
+      render json: { error: 'Not Authorized' }, status: 401
+    end
   end
 
   def destroy
-    @category = Category.find(params[:id])
-    return render json: @category.errors, status: :unprocessable_entity unless @category.destroy
+    if @current_user.role == 'admin'
+      @category = Category.find(params[:id])
+      return render json: @category.errors, status: :unprocessable_entity unless @category.destroy
 
-    render json: { status: :no_content, message: 'Category deleted successfully.', body: nil }, status: :ok
+      render json: { status: :no_content, message: 'Category deleted successfully.', body: nil }, status: :ok
+    else
+      render json: { error: 'Not Authorized' }, status: 401
+    end
   end
 end
